refactor: Moved correlation plot tab
